ADDS finalize pool response & move queries from staking data service
